Refactor: 중복되는 코드 정리 #23
