+++ src/components/MonthlyChart/MonthlyChart.jsx
@@ -10,15 +10,6 @@ import * as S from "./MonthlyChart.style";
 
 import { useState } from "react";
 
-// 날짜 포맷팅 함수
-const formatDate = (date) => {
-  let year = date.getFullYear(); // 년
-  let month = String(date.getMonth() + 1).padStart(2, "0"); // 월
-  let day = String(date.getDate()).padStart(2, "0"); // 일
-
-  return `${year}-${month}-${day}`; // 년-월-일 형태로 반환 (문자열 형태, 더미데이터에서 비교를 위해)
-};
-
 const MonthlyChart = ({ date }) => {
   const [pieIcon, setPieIcon] = useState(pieActive); // 원형 그래프
   const [barIcon, setBarIcon] = useState(barUnactive); // 막대 그래프
@@ -27,14 +18,6 @@ const MonthlyChart = ({ date }) => {
     setPieIcon((prev) => (prev === pieActive ? pieUnactive : pieActive));
     setBarIcon((prev) => (prev === barUnactive ? barActive : barUnactive));
   };
-
-  //const [date_, setDate] = useState(formatDate(new Date(date)));
-  //console.log(date_);
-
-  // const handleDateChange = (newDate) => {
-  //   const formattedDate = typeof newDate === "object" && newDate instanceof Date ? formatDate(newDate) : newDate; // 포맷 (문자열로)
-  //   setDate(formattedDate); // 새로운 날짜 상태 업데이트
-  // };
 
   return (
     <S.ChartDiv>
